refactor: user-details module remediation (PR-4)
