--- .changie.yaml
@@ -1,19 +1,0 @@
-changesDir: .changes
-unreleasedDir: unreleased
-headerPath: header.tpl.md
-changelogPath: CHANGELOG.md
-versionExt: md
-versionFormat: '## {{.Version}} - {{.Time.Format "2006-01-02"}}'
-kindFormat: '### {{.Kind}}'
-changeFormat: '* {{.Body}}'
-kinds:
-  - label: Added
-  - label: Changed
-  - label: Deprecated
-  - label: Removed
-  - label: Fixed
-  - label: Security
-newlines:
-  afterChangelogHeader: 1
-  beforeChangelogVersion: 1
-  endOfVersion: 1
