+++ src/modules/user-details/http.ts
@@ -1,9 +1,12 @@
 import * as handlers from '@/modules/user-details/handlers';
 import * as routes from '@/modules/user-details/routes';
+import { injectAbility } from '@/shared/middleware/inject-ability';
 import { createHonoApp } from '@/shared/router/factory';
 import { registerOpenApiRoutes } from '@/shared/router/openapi-docs';
 
 const userDetailsApp = createHonoApp();
+
+userDetailsApp.use('*', injectAbility());
 
 // User Details (Note: No POST/create - clients are created via intake or invitation flows)
 userDetailsApp.openapi(routes.listUserDetailsRoute, handlers.listUserDetailsHandler);
@@ -15,7 +18,6 @@ userDetailsApp.openapi(routes.listUserDetailsMemosRoute, handlers.listUserDetail
 userDetailsApp.openapi(routes.createUserDetailMemoRoute, handlers.createUserDetailMemoHandler);
 userDetailsApp.openapi(routes.updateUserDetailsMemoRoute, handlers.updateUserDetailMemoHandler);
 userDetailsApp.openapi(routes.deleteUserDetailsMemoRoute, handlers.deleteUserDetailMemoHandler);
-
 
 // Register routes for OpenAPI documentation extraction
 registerOpenApiRoutes(userDetailsApp, routes);
